SERVER-12819 Reject write cmds where batch size > maxWriteBatchSize

+++ src/mongo/s/write_ops/batch_upconvert.cpp
@@ -88,7 +88,7 @@ namespace mongo {
                     break;
                 }
             }
-            while ( static_cast<int>( docs.size() ) < BatchedCommandRequest::kMaxWriteBatchSize
+            while ( docs.size() < BatchedCommandRequest::kMaxWriteBatchSize
                     && dbMsg.moreJSObjs() );
 
             dassert( !docs.empty() );
